requirements.txt
garminconnect>=0.2.25
requests>=2.31.0
python-dateutil>=2.8.2

